Refractor Initializer (#683)
* initial commit

* Refractor InitializerFunctions.scala

* Refractor InitializerFunctions.scala

* Change Scala Sources Name

* Refractor InitializerFunctions.scala

* Refractor InitializerFunctions.scala

* Added Initializerv2.scala

* Added Initializerv2.scala

* Changed as per Sriram comment

* Changed as per Sriram comment

* dropped Initializer Deprecated

---------
