+++ src/main/scala/com/databricks/labs/overwatch/utils/Config.scala
@@ -202,7 +202,6 @@ class Config() {
     this
   }
 
-
   private[overwatch] def setCloudProvider(value: String): this.type = {
     _cloudProvider = value
     this
@@ -303,7 +302,7 @@ class Config() {
     this
   }
 
-  private def setApiEnv(value: ApiEnv): this.type = {
+  def setApiEnv(value: ApiEnv): this.type = {
     _apiEnv = value
     this
   }
@@ -317,6 +316,16 @@ class Config() {
 
   def setInputConfig(value: OverwatchParams): this.type = {
     _inputConfig = value
+    this
+  }
+
+  private[overwatch] def setWorkspaceURL(value: String): this.type = {
+    _workspaceUrl = value
+    this
+  }
+
+  private[overwatch] def setTokenType(value: String): this.type = {
+    _tokenType = value
     this
   }
 
@@ -388,8 +397,63 @@ class Config() {
     }
   }
 
-
-
+  private[overwatch] def buildApiEnv(tokenSecret: Option[TokenSecret], apiEnvConfig: Option[ApiEnvConfig]): ApiEnv = {
+    var rawToken = ""
+    var scope = ""
+    var key = ""
+    if (isMultiworkspaceDeployment && apiUrl.nonEmpty) {
+      setWorkspaceURL(apiUrl.get)
+      logger.log(Level.INFO, "Multiworkspace Deployment setting the workspaceURL :" + _workspaceUrl)
+    } else {
+      setWorkspaceURL(dbutils.notebook.getContext().apiUrl.get)
+    }
+    setCloudProvider(if (_workspaceUrl.toLowerCase().contains("azure")) "azure" else "aws")
+    try {
+      // Token secrets not supported in local testing
+      if (tokenSecret.nonEmpty) { // not local testing and secret passed
+        scope = tokenSecret.get.scope
+        key = tokenSecret.get.key
+        rawToken = dbutils.secrets.get(scope, key)
+        val authMessage = s"Valid Secret Identified: Executing with token located in secret, $scope : $key"
+        logger.log(Level.INFO, authMessage)
+        setTokenType("Secret")
+      } else { // Use default token for job owner
+        rawToken = dbutils.notebook.getContext().apiToken.get
+        val authMessage = "No secret parameters provided: attempting to continue with job owner's token."
+        logger.log(Level.WARN, authMessage)
+        println(authMessage)
+        setTokenType("Owner")
+      }
+      if (!rawToken.matches("^(dapi|dkea)[a-zA-Z0-9-]*$")) throw new BadConfigException(s"contents of secret " +
+        s"at scope:key $scope:$key is not in a valid format. Please validate the contents of your secret. It must be " +
+        s"a user access token. It should start with 'dapi' ")
+      val derivedApiEnvConfig = apiEnvConfig.getOrElse(ApiEnvConfig())
+      val derivedApiProxy = derivedApiEnvConfig.apiProxyConfig.getOrElse(ApiProxyConfig())
+      ApiEnv(
+        isLocalTesting,
+        workspaceURL,
+        rawToken,
+        packageVersion,
+        derivedApiEnvConfig.successBatchSize,
+        derivedApiEnvConfig.errorBatchSize,
+        runID,
+        derivedApiEnvConfig.enableUnsafeSSL,
+        derivedApiEnvConfig.threadPoolSize,
+        derivedApiEnvConfig.apiWaitingTime,
+        derivedApiProxy.proxyHost,
+        derivedApiProxy.proxyPort,
+        derivedApiProxy.proxyUserName,
+        derivedApiProxy.proxyPasswordScope,
+        derivedApiProxy.proxyPasswordKey
+      )
+    } catch {
+      case e: IllegalArgumentException if e.getMessage.toLowerCase.contains("secret does not exist with scope") =>
+        throw new BadConfigException(e.getMessage, failPipeline = true)
+      case e: Throwable =>
+        logger.log(Level.FATAL, "No valid credentials and/or Databricks URI", e)
+        throw new BadConfigException(e.getMessage, failPipeline = true)
+    }
+  }
 
   /**
    * Set Overwatch DB and location
